+++ R/chemdose_ph.R
@@ -27,14 +27,16 @@
 #' @param hcl Amount of hydrochloric acid added in mg/L: HCl -> H + Cl
 #' @param h2so4 Amount of sulfuric acid added in mg/L: H2SO4 -> 2H + SO4
 #' @param h3po4 Amount of phosphoric acid added in mg/L: H3PO4 -> 3H + PO4
+#' @param hno3 Amount of nitric acid added in mg/L: HNO3 -> H + NO3
 #' @param co2 Amount of carbon dioxide added in mg/L: CO2 (gas) + H2O -> H2CO3*
 #' @param naoh Amount of caustic added in mg/L: NaOH -> Na + OH
 #' @param caoh2 Amount of lime added in mg/L: Ca(OH)2 -> Ca + 2OH
-#' @param mgoh2  Amount of magneisum hydroxide added in mg/L: Mg(OH)2 -> Mg + 2OH
+#' @param mgoh2  Amount of magnesium hydroxide added in mg/L: Mg(OH)2 -> Mg + 2OH
 #' @param na2co3 Amount of soda ash added in mg/L: Na2CO3 -> 2Na + CO3
 #' @param nahco3 Amount of sodium bicarbonate added in mg/L: NaHCO3 -> Na + H + CO3
 #' @param caco3 Amount of calcium carbonate added (or removed) in mg/L: CaCO3 -> Ca + CO3
 #' @param caso4 Amount of calcium sulfate added (for post-RO condition) in mg/L: CaSO4 -> Ca + SO4
+#' @param caocl2 Amount of Calcium hypochlorite added in mg/L as Cl2: CaOCl2 -> Ca + 2OCl
 #' @param cacl2 Amount of calcium chloride added in mg/L: CaCl2 -> Ca2+ + 2Cl-
 #' @param cl2 Amount of chlorine gas added in mg/L as Cl2: Cl2(g) + H2O -> HOCl + H + Cl
 #' @param naocl Amount of sodium hypochlorite added in mg/L as Cl2: NaOCl -> Na + OCl
@@ -44,6 +46,9 @@
 #' @param ferricchloride Amount of ferric Chloride added in mg/L: FeCl3 + 3HCO3 -> Fe(OH)3(am) + 3Cl + 3CO2
 #' @param ferricsulfate Amount of ferric sulfate added in mg/L: Fe2(SO4)3*8.8H2O + 6HCO3 -> 2Fe(OH)3(am) + 3SO4 + 8.8H2O + 6CO2
 #' @param ach Amount of aluminum chlorohydrate added in mg/L: Al2(OH)5Cl*2H2O + HCO3 -> 2Al(OH)3(am) + Cl + 2H2O + CO2
+#' @param kmno4 Amount of potassium permanganate added in mg/L: KMnO4 -> K + MnO4
+#' @param naf Amount of sodium fluoride added in mg/L: NaF -> Na + F
+#' @param na3po4 Amount of trisodium phosphate added in mg/L: Na3PO4 -> 3Na + PO4
 #' @param softening_correction Set to TRUE to correct post-softening pH (caco3 must be < 0). Default is FALSE. Based on WTP model equation 5-62
 #'
 #' @seealso \code{\link{define_water}}, \code{\link{convert_units}}
@@ -75,11 +80,12 @@
 #'
 #' @returns `chemdose_ph` returns a water class object with updated pH, alkalinity, and ions post-chemical addition.
 #'
-chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
+chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, hno3 = 0, co2 = 0,
                         naoh = 0, caoh2 = 0, mgoh2 = 0,
-                        na2co3 = 0, nahco3 = 0, caco3 = 0, caso4 = 0, cacl2 = 0,
+                        na2co3 = 0, nahco3 = 0, caco3 = 0, caso4 = 0, caocl2 = 0, cacl2 = 0,
                         cl2 = 0, naocl = 0, nh4oh = 0, nh42so4 = 0,
                         alum = 0, ferricchloride = 0, ferricsulfate = 0, ach = 0,
+                        kmno4 = 0, naf = 0, na3po4 = 0,
                         softening_correction = FALSE) {
   if ((cacl2 > 0 | cl2 > 0 | naocl > 0) & (nh4oh > 0 | nh42so4 > 0)) {
     warning("Both chlorine- and ammonia-based chemicals were dosed and may form chloramines.\nUse chemdose_chloramine for breakpoint caclulations.")
@@ -102,6 +108,8 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   h2so4 <- convert_units(h2so4, "h2so4")
   # Phosphoric acid (H3PO4) dose
   h3po4 <- convert_units(h3po4, "h3po4")
+  # Nitric acid (HNO3) dose
+  hno3 <- convert_units(hno3, "hno3")
   # Carbon dioxide
   co2 <- convert_units(co2, "co2")
 
@@ -120,8 +128,12 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   cacl2 <- convert_units(cacl2, "cacl2")
   # Chlorine gas (Cl2)
   cl2 <- convert_units(cl2, "cl2")
+  
   # Sodium hypochlorite (NaOCl) as Cl2
   naocl <- convert_units(naocl, "cl2")
+  
+  # Calcium hypochlorite (CaOCl2) 
+  caocl2 <- convert_units(caocl2, "cl2")
 
   # CaCO3
   caco3 <- convert_units(caco3, "caco3")
@@ -142,16 +154,24 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   ferricsulfate <- convert_units(ferricsulfate, "ferricsulfate")
   # ACH
   ach <- convert_units(ach, "ach")
+  
+  # Potassium permanganate (KMnO4) dose
+  kmno4 <- convert_units(kmno4, "kmno4")
+  # Sodium fluoride (NaF) dose
+  naf <- convert_units(naf, "naf")
+  # Trisodium phosphate (Na3PO4) dose
+  na3po4 <- convert_units(na3po4, "na3po4")
+  
 
   #### CALCULATE NEW ION BALANCE FROM ALL CHEMICAL ADDITIONS ####
   dosed_water <- water
 
   # Total sodium
-  na_dose <- naoh + 2 * na2co3 + nahco3 + naocl
+  na_dose <- naoh + 2 * na2co3 + nahco3 + naocl + naf + 3 * na3po4
   dosed_water@na <- water@na + na_dose
 
   # Total calcium
-  ca_dose <- caoh2 + cacl2 + caco3 + caso4
+  ca_dose <- caoh2 + caocl2 / 2 + cacl2 + caco3 + caso4
   dosed_water@ca <- water@ca + ca_dose
 
   # Total magnesium
@@ -159,8 +179,12 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   dosed_water@mg <- water@mg + mg_dose
 
   # Total potassium
-  k_dose <- 0
+  k_dose <- kmno4
   dosed_water@k <- water@k + k_dose
+  
+  # Total permanganate
+  mno4_dose <- kmno4
+  dosed_water@mno4 <- water@mno4 + mno4_dose
 
   # Total chloride
   cl_dose <- hcl + cl2 + 2 * cacl2 + 3 * ferricchloride + ach
@@ -171,11 +195,11 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   dosed_water@so4 <- water@so4 + so4_dose
 
   # Total phosphate
-  po4_dose <- h3po4
+  po4_dose <- h3po4 + na3po4
   dosed_water@tot_po4 <- water@tot_po4 + po4_dose
 
   # Total hypochlorite
-  ocl_dose <- cl2 + naocl
+  ocl_dose <- cl2 + naocl + caocl2
   dosed_water@free_chlorine <- water@free_chlorine + ocl_dose
 
   # Total ammonia
@@ -193,7 +217,7 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
     convert_units(ca_dose, "ca", "M", "mg/L") + convert_units(mg_dose, "mg", "M", "mg/L") +
     convert_units(co3_dose - co2, "co3", "M", "mg/L") + convert_units(po4_dose, "po4", "M", "mg/L") +
     convert_units(so4_dose, "so4", "M", "mg/L") + convert_units(ocl_dose, "ocl", "M", "mg/L") +
-    convert_units(nh4_dose, "nh4", "M", "mg/L")
+    convert_units(nh4_dose, "nh4", "M", "mg/L") + convert_units(mno4_dose, "mno4", "M", "mg/L")
   if (!is.na(dosed_water@tds) & dosed_water@tds < 0) {
     warning("Calculated TDS after chemical removal < 0. TDS and ionic strength will be set to 0.")
     dosed_water@tds <- 0
@@ -202,7 +226,7 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
   dosed_water@cond <- correlate_ionicstrength(dosed_water@tds, from = "tds", to = "cond")
 
   # Calculate new pH, H+ and OH- concentrations
-  ph <- solve_ph(dosed_water, so4_dose = so4_dose, na_dose = na_dose, ca_dose = ca_dose, mg_dose = mg_dose, cl_dose = cl_dose)
+  ph <- solve_ph(dosed_water, so4_dose = so4_dose, na_dose = na_dose, ca_dose = ca_dose, mg_dose = mg_dose, cl_dose = cl_dose, mno4_dose = mno4_dose)
 
   if (softening_correction == TRUE & caco3 < 0) {
     ph_corrected <- (ph - 1.86) / 0.71 # WTP Model eq 5-62
@@ -296,17 +320,18 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, co2 = 0,
 #' @returns `chemdose_ph_chain` returns a data frame containing a water class column with updated pH, alkalinity, and ions post-chemical addition.
 
 chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = "dosed_chem_water",
-                              hcl = "use_col", h2so4 = "use_col", h3po4 = "use_col", co2 = "use_col", naoh = "use_col",
+                              hcl = "use_col", h2so4 = "use_col", h3po4 = "use_col", hno3 = "use_col", co2 = "use_col", naoh = "use_col",
                               na2co3 = "use_col", nahco3 = "use_col", caoh2 = "use_col", mgoh2 = "use_col",
-                              cacl2 = "use_col", cl2 = "use_col", naocl = "use_col",
-                              nh4oh = "use_col", nh42so4 = "use_col",
+                              caocl2 = "use_col", cacl2 = "use_col", cl2 = "use_col", naocl = "use_col",
+                              nh4oh = "use_col", nh42so4 = "use_col", caco3 = "use_col", caso4 = "use_col",
                               alum = "use_col", ferricchloride = "use_col", ferricsulfate = "use_col", ach = "use_col",
-                              caco3 = "use_col", caso4 = "use_col", softening_correction = "use_col", na_to_zero = TRUE) {
+                              kmno4 = "use_col", naf = "use_col", na3po4 = "use_col", softening_correction = "use_col", na_to_zero = TRUE) {
   validate_water_helpers(df, input_water)
   # This allows for the function to process unquoted column names without erroring
   hcl <- tryCatch(hcl, error = function(e) enquo(hcl))
   h2so4 <- tryCatch(h2so4, error = function(e) enquo(h2so4))
   h3po4 <- tryCatch(h3po4, error = function(e) enquo(h3po4))
+  hno3 <- tryCatch(hno3, error = function(e) enquo(hno3))
   co2 <- tryCatch(co2, error = function(e) enquo(co2))
   naoh <- tryCatch(naoh, error = function(e) enquo(naoh))
 
@@ -315,6 +340,7 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
   caoh2 <- tryCatch(caoh2, error = function(e) enquo(caoh2))
   mgoh2 <- tryCatch(mgoh2, error = function(e) enquo(mgoh2))
 
+  caocl2 <- tryCatch(caocl2, error = function(e) enquo(caocl2))
   cacl2 <- tryCatch(cacl2, error = function(e) enquo(cacl2))
   cl2 <- tryCatch(cl2, error = function(e) enquo(cl2))
   naocl <- tryCatch(naocl, error = function(e) enquo(naocl))
@@ -328,16 +354,21 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
   ach <- tryCatch(ach, error = function(e) enquo(ach))
   caco3 <- tryCatch(caco3, error = function(e) enquo(caco3))
   caso4 <- tryCatch(caso4, error = function(e) enquo(caso4))
+  
+  kmno4 <- tryCatch(kmno4, error = function(e) enquo(kmno4))
+  naf <- tryCatch(naf, error = function(e) enquo(naf))
+  na3po4 <- tryCatch(na3po4, error = function(e) enquo(na3po4))
 
   softening_correction <- tryCatch(softening_correction, error = function(e) enquo(softening_correction))
 
   # This returns a dataframe of the input arguments and the correct column names for the others
   arguments <- construct_helper(df, all_args = list(
-    "hcl" = hcl, "h2so4" = h2so4, "h3po4" = h3po4, "co2" = co2, "naoh" = naoh,
+    "hcl" = hcl, "h2so4" = h2so4, "h3po4" = h3po4, "hno3" = hno3, "co2" = co2, "naoh" = naoh,
     "na2co3" = na2co3, "nahco3" = nahco3, "caoh2" = caoh2, "mgoh2" = mgoh2,
-    "cacl2" = cacl2, "cl2" = cl2, "naocl" = naocl,
-    "nh4oh" = nh4oh, "nh42so4" = nh42so4,
-    "alum" = alum, "ferricchloride" = ferricchloride, "ferricsulfate" = ferricsulfate, "ach" = ach, "caco3" = caco3, "caso4" = caso4,
+    "caocl2" = caocl2, "cacl2" = cacl2, "cl2" = cl2, "naocl" = naocl,
+    "nh4oh" = nh4oh, "nh42so4" = nh42so4, "caco3" = caco3, "caso4" = caso4,
+    "alum" = alum, "ferricchloride" = ferricchloride, "ferricsulfate" = ferricsulfate, "ach" = ach,
+    "kmno4" = kmno4, "naf" = naf, "na3po4" = na3po4,
     "softening_correction" = softening_correction
   ))
   final_names <- arguments$final_names
@@ -360,12 +391,14 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
         hcl = if (final_names$hcl %in% names(.)) !!sym(final_names$hcl) else rep(0, nrow(.)),
         h2so4 = if (final_names$h2so4 %in% names(.)) !!sym(final_names$h2so4) else rep(0, nrow(.)),
         h3po4 = if (final_names$h3po4 %in% names(.)) !!sym(final_names$h3po4) else rep(0, nrow(.)),
+        hno3 = if (final_names$hno3 %in% names(.)) !!sym(final_names$hno3) else rep(0, nrow(.)),
         co2 = if (final_names$co2 %in% names(.)) !!sym(final_names$co2) else rep(0, nrow(.)),
         naoh = if (final_names$naoh %in% names(.)) !!sym(final_names$naoh) else rep(0, nrow(.)),
         na2co3 = if (final_names$na2co3 %in% names(.)) !!sym(final_names$na2co3) else rep(0, nrow(.)),
         nahco3 = if (final_names$nahco3 %in% names(.)) !!sym(final_names$nahco3) else rep(0, nrow(.)),
         caoh2 = if (final_names$caoh2 %in% names(.)) !!sym(final_names$caoh2) else rep(0, nrow(.)),
         mgoh2 = if (final_names$mgoh2 %in% names(.)) !!sym(final_names$mgoh2) else rep(0, nrow(.)),
+        caocl2 = if (final_names$caocl2 %in% names(.)) !!sym(final_names$caocl2) else rep(0, nrow(.)),
         cacl2 = if (final_names$cacl2 %in% names(.)) !!sym(final_names$cacl2) else rep(0, nrow(.)),
         cl2 = if (final_names$cl2 %in% names(.)) !!sym(final_names$cl2) else rep(0, nrow(.)),
         naocl = if (final_names$naocl %in% names(.)) !!sym(final_names$naocl) else rep(0, nrow(.)),
@@ -377,6 +410,9 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
         ach = if (final_names$ach %in% names(.)) !!sym(final_names$ach) else rep(0, nrow(.)),
         caco3 = if (final_names$caco3 %in% names(.)) !!sym(final_names$caco3) else rep(0, nrow(.)),
         caso4 = if (final_names$caso4 %in% names(.)) !!sym(final_names$caso4) else rep(0, nrow(.)),
+        kmno4 = if (final_names$kmno4 %in% names(.)) !!sym(final_names$kmno4) else rep(0, nrow(.)),
+        naf = if (final_names$naf %in% names(.)) !!sym(final_names$naf) else rep(0, nrow(.)),
+        na3po4 = if (final_names$na3po4 %in% names(.)) !!sym(final_names$na3po4) else rep(0, nrow(.)),
         softening_correction = if (final_names$softening_correction %in% names(.)) !!sym(final_names$softening_correction) else rep(FALSE, nrow(.))
       ),
       chemdose_ph
@@ -415,17 +451,19 @@ chemdose_ph_chain <- function(df, input_water = "defined_water", output_water = 
 #'
 
 chemdose_ph_once <- function(df, input_water = "defined_water",
-                             hcl = "use_col", h2so4 = "use_col", h3po4 = "use_col", co2 = "use_col", naoh = "use_col",
+                             hcl = "use_col", h2so4 = "use_col", h3po4 = "use_col", hno3 = "use_col", co2 = "use_col", naoh = "use_col",
                              na2co3 = "use_col", nahco3 = "use_col", caoh2 = "use_col", mgoh2 = "use_col",
-                             cl2 = "use_col", naocl = "use_col", nh4oh = "use_col", nh42so4 = "use_col",
+                             caocl2 = "use_col", cacl2 = "use_col", cl2 = "use_col", naocl = "use_col",
+                             nh4oh = "use_col", nh42so4 = "use_col", caco3 = "use_col", caso4 = "use_col",
                              alum = "use_col", ferricchloride = "use_col", ferricsulfate = "use_col", ach = "use_col",
-                             caco3 = "use_col") {
+                             kmno4 = "use_col", naf = "use_col", na3po4 = "use_col") {
   dose_chem <- dosed_chem_water <- temp <- tds <- estimated <- NULL # Quiet RCMD check global variable note
   
   # This allows for the function to process unquoted column names without erroring
   hcl <- tryCatch(hcl, error = function(e) enquo(hcl))
   h2so4 <- tryCatch(h2so4, error = function(e) enquo(h2so4))
   h3po4 <- tryCatch(h3po4, error = function(e) enquo(h3po4))
+  hno3 <- tryCatch(hno3, error = function(e) enquo(hno3))
   co2 <- tryCatch(co2, error = function(e) enquo(co2))
   naoh <- tryCatch(naoh, error = function(e) enquo(naoh))
   
@@ -434,8 +472,11 @@ chemdose_ph_once <- function(df, input_water = "defined_water",
   caoh2 <- tryCatch(caoh2, error = function(e) enquo(caoh2))
   mgoh2 <- tryCatch(mgoh2, error = function(e) enquo(mgoh2))
   
+  caocl2 <- tryCatch(caocl2, error = function(e) enquo(caocl2))
+  cacl2 <- tryCatch(cacl2, error = function(e) enquo(cacl2))
   cl2 <- tryCatch(cl2, error = function(e) enquo(cl2))
   naocl <- tryCatch(naocl, error = function(e) enquo(naocl))
+  
   nh4oh <- tryCatch(nh4oh, error = function(e) enquo(nh4oh))
   nh42so4 <- tryCatch(nh42so4, error = function(e) enquo(nh42so4))
   
@@ -444,14 +485,21 @@ chemdose_ph_once <- function(df, input_water = "defined_water",
   ferricsulfate <- tryCatch(ferricsulfate, error = function(e) enquo(ferricsulfate))
   ach <- tryCatch(ach, error = function(e) enquo(ach))
   caco3 <- tryCatch(caco3, error = function(e) enquo(caco3))
+  caso4 <- tryCatch(caso4, error = function(e) enquo(caso4))
+  
+  kmno4 <- tryCatch(kmno4, error = function(e) enquo(kmno4))
+  naf <- tryCatch(naf, error = function(e) enquo(naf))
+  na3po4 <- tryCatch(na3po4, error = function(e) enquo(na3po4))
   
   output <- df %>%
     chemdose_ph_chain(
       input_water = input_water, output_water = "dosed_chem_water",
-      hcl, h2so4, h3po4, co2, naoh,
+      hcl, h2so4, h3po4, hno3, co2, naoh,
       na2co3, nahco3, caoh2, mgoh2,
-      cl2, naocl, nh4oh, nh42so4,
-      alum, ferricchloride, ferricsulfate, ach, caco3
+      caocl2, cacl2, cl2, naocl,
+      nh4oh, nh42so4, caco3, caso4,
+      alum, ferricchloride, ferricsulfate, ach,
+      kmno4, naf, na3po4
     ) %>%
     mutate(dose_chem = furrr::future_map(dosed_chem_water, convert_water)) %>%
   unnest(dose_chem) %>%